Dynamically reweight estimator using update frequency data
Prior to this commit, the estimator uses a double exponentially
weighted average with a constant weight parameter set such that
90% of the estimate comes from the most recent 15 seconds of tick
data.

This does not work well for all workloads. For instance, if 15
seconds have elapsed since the last tick, indicatif will
*correctly* note that progress has been stalled - 90% of the
estimate will be from a period where zero progress has been made.
For e.g. most file transfers this is pretty reasonable, because you
*want* the estimate to be responsive to the the connection dropping
which should rapidly send the ETA to very high values. For other
use cases sending the estimate through the roof until more data
comes in is unacceptable.

Users of indicatif with unpredictable ticks should seriously
consider whether a steady_tick is appropriate for their application,
as disabling the steady tick eliminates the rapidly climbing
estimate in the case of stalled progress.

One potential solution would be to make the EWA tuning parameter
adjustable by consumers of the indicatif library. However, it was
decided that adding an additional parameter hurts the ergonomics
of the library.

Instead, this commit fudges the weight as we go. This breaks the
clean mathematics of the EWA, but may be a reasonable pragmatic
choice. Here, we generate an unbiased estimate of the maximum
time between ticks, and set the weight to a multiple (6.6x) this
value. This value is used whenever it exceeds the 15 second default
weight. In practice, this means that an ETA will approximately
double when twice the estimated maximum time between ticks has
elapsed without an update occurring.

The exact value could be debated, there's nothing inherently
right or wrong about 6.6.

Note that some users might potentially object to the estimation
increasing over time when no progress is made. This is
fundamentally based on a mistake. Sending a tick is equivalent to
requesting a new point estimate of the completion time from
indicatif, and this means we should return the most accurate
estimate we can. If time has elapsed with no new progress, then
this estimate *must* rise. Users who object to this should send
their own ticks rather than using steady_tick, since the whole
point of the latter is to get continual updates.

Others might even insist that the ETA should continue to go *down*
even though no explicit progress has been made, similar to the
classic Windows file copy dialog. I think it's a mistake to do
this in a general purpose progress-bar library, because there are
(obviously) many pathological cases where stalls are not
adequately reported or are reported in a frustrating fashion. Users
who want this kind of progress bar would probably be better off
writing their own.